% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Conformal_select}
\alias{Conformal_select}
\title{Conformalized Selection}
\usage{
Conformal_select(cal.score, test.score, q)
}
\arguments{
\item{cal.score}{Vector of scores V_i = V(X_i,Y_i) for calibration data}

\item{test.score}{Vector of scores hat{V}\emph{{n+j} = V(X}{n+j}, c_{n+j}) for test data}

\item{q}{Nominal FDR level, default at 0.1}
}
\value{
A vector of indices in the test data that are selected as 'promising candidates'
}
\description{
This function implements the Conformalized Selection algorithm (WCS without weights, for i.i.d. or exchangeable data). It takes evaluated scores for the calibration and test data as input, and returns the index set of the selected units with a specified FDR nominal level.
}
