% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weighted_CS}
\alias{weighted_CS}
\title{Weighted Conformalized Selection}
\usage{
weighted_CS(
  cal.score,
  test.score,
  cal.weight,
  test.weight,
  q = 0.1,
  rand = "hete"
)
}
\arguments{
\item{cal.score}{Vector of scores V_i = V(X_i,Y_i) for calibration data}

\item{test.score}{Vector of scores hat{V}\emph{{n+j} = V(X}{n+j}, c_{n+j}) for test data}

\item{cal.weight}{Vector of weights W_i = W(X_i) for cailbration data}

\item{test.weight}{Vector of weights W_{n+j} = W(X_{n+j}) for test data}

\item{q}{Nominal FDR level, default at 0.1}

\item{rand}{Method for random pruning, can be 'hete' for heterogeneous pruning, 'homo' for homogeneous pruning, and 'dtm' for deterministic pruning}
}
\value{
A vector of indices in the test data that are selected as 'promising candidates'
}
\description{
This function implements the Weighted Conformalized Selection algorithm. It takes evaluated scores and weights for the calibration and test data as input, and returns the index set of the selected units with a specified FDR nominal level.
}
